fix: rename Avatar component too UserAvatar to differentiate between that and mui Avatar
